moving function to ivadomed/maths
